+++ bin/fzf-preview.sh
@@ -4,8 +4,9 @@
 # image in the preview window of fzf.
 #
 # Dependencies:
-# - https://github.com/hpjansson/chafa
 # - https://github.com/sharkdp/bat
+# - https://github.com/hpjansson/chafa
+# - https://iterm2.com/utilities/imgcat
 
 if [[ $# -ne 1 ]]; then
   >&2 echo "usage: $0 FILENAME"
@@ -44,6 +45,7 @@ elif ! [[ $KITTY_WINDOW_ID ]] && (( FZF_PREVIEW_TOP + FZF_PREVIEW_LINES == $(stt
   dim=${FZF_PREVIEW_COLUMNS}x$((FZF_PREVIEW_LINES - 1))
 fi
 
+# 1. Use kitty icat on kitty terminal
 if [[ $KITTY_WINDOW_ID ]]; then
   # 1. 'memory' is the fastest option but if you want the image to be scrollable,
   #    you have to use 'stream'.
@@ -52,10 +54,21 @@ if [[ $KITTY_WINDOW_ID ]]; then
   #    This confuses fzf and makes it render scroll offset indicator.
   #    So we remove the last line and append the reset code to its previous line.
   kitty icat --clear --transfer-mode=memory --stdin=no --place="$dim@0x0" "$file" | sed '$d' | sed $'$s/$/\e[m/'
+
+# 2. Use chafa with Sixel output
 elif command -v chafa > /dev/null; then
   chafa -f sixel -s "$dim" "$file"
   # Add a new line character so that fzf can display multiple images in the preview window
   echo
+
+# 3. If chafa is not found but imgcat is available, use it on iTerm2
+elif command -v imgcat > /dev/null; then
+  # NOTE: We should use https://iterm2.com/utilities/it2check to check if the
+  # user is running iTerm2. But for the sake of simplicty, we just assume
+  # that's the case here.
+  imgcat -W "${dim%%x*}" -H "${dim##*x}" "$file"
+
+# 4. Cannot find any suitable method to preview the image
 else
   file "$file"
 fi